# digital-laundry-backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "Laundry Management API", "version": "1.0.0" },
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/auth/register/staff": {
      "post": {
        "summary": "Register staff",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fullName", "email", "staffCode", "password"],
                "properties": {
                  "fullName": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "staffCode": { "type": "string" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registered. Verify email." },
          "400": { "description": "Invalid" }
        }
      }
    },
...
    "/api/admin/hostels": {
      "post": {
        "summary": "Create hostel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hostelName", "hostelCode", "status"],
                "properties": {
                  "hostelName": { "type": "string" },
                  "hostelCode": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Hostel" },
          "409": { "description": "hostelCode exists" }
        }
      },
      "get": {
        "summary": "List hostels",
        "security": [],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
          }
        ],
        "responses": { "200": { "description": "Hostels" } }
      }
    },
    "/api/admin/hostels/{id}": {
      "put": {
        "summary": "Update hostel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hostelName", "hostelCode", "status"],
                "properties": {
                  "hostelName": { "type": "string" },
                  "hostelCode": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Hostel" },
          "404": { "description": "Not found" },
          "409": { "description": "hostelCode exists" }
        }
      },
      "delete": {
        "summary": "Delete hostel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" },
          "409": { "description": "In use" }
        }
      }
    },
    "/api/admin/staff-codes": {
      "post": {
        "summary": "Create staff code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "status"],
                "properties": {
                  "code": { "type": "string" },
                  "staffNameHint": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "StaffCode" },
          "409": { "description": "code exists" }
        }
      },
      "get": {
        "summary": "List staff codes",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
          },
          { "name": "claimed", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Staff codes" } }
      }
    },
    "/api/admin/staff-codes/{code}": {
      "put": {
        "summary": "Update staff code",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "status"],
                "properties": {
                  "code": { "type": "string" },
                  "staffNameHint": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "StaffCode" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete staff code",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" },
          "409": { "description": "In use" }
        }
      }
    },
    "/api/admin/students": {
      "get": {
        "summary": "List students",
        "parameters": [
          {
            "name": "hostelCode",
            "in": "query",
            "schema": { "type": "string" }
          },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/admin/staff": {
      "get": {
        "summary": "List staff",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/admin/orders": {
      "get": {
        "summary": "List orders",
        "parameters": [
          {
            "name": "hostelCode",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "uniqueLaundryId",
            "in": "query",
            "schema": { "type": "string" }
          },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": { "type": "string", "format": "date" }
          },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Create order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderDate", "items"],
                "properties": {
                  "orderDate": { "type": "string" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["itemType", "quantity"],
                      "properties": {
                        "itemType": { "type": "string" },
                        "quantity": { "type": "integer" },
                        "remark": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order created" },
          "400": { "description": "Validation" }
        }
      },
      "get": {
        "summary": "List orders",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "sort", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paged" } }
      }
    },
...
    "/api/dashboard/admin": {
      "get": {
        "summary": "Admin dashboard counts",
        "parameters": [
          {
            "name": "hostelCode",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "Counts" } }
      }
    },
...
  }
}